requirements.txt
# Core dependencies
pyyaml>=6.0
pydantic>=2.0
numpy>=1.24.0
httpx>=0.24.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21.0

# LLM (optional)
openai>=1.0.0
anthropic>=0.18.0

# Development
black>=23.0
ruff>=0.1.0
mypy>=1.0
